requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.79
botocore==1.34.79
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
jmespath==1.0.1
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==1.26.18
